just some additions and cleanings

+++ backend/db/repository/blog.py
@@ -24,7 +24,9 @@ def list_blogs(db: Session):
 def update_blog(id: int, blog: UpdateBlog, author_id: int, db: Session):
     blog_in_db = db.query(Blog).filter(Blog.id == id).first()
     if not blog_in_db:
-        return
+        return {"error": f"Could not find blog with id {id}"}
+    if not blog_in_db.author_id == author_id:
+        return {"error": "Only author can update the blog"}
     blog_in_db.title = blog.title
     blog_in_db.content = blog.content
     db.add(blog_in_db)
@@ -37,7 +39,9 @@ def delete_blog(id: int, author_id: int, db: Session):
     #we are working with the reference of the blog object and deleting it
     #because the method .delete() does not defined in this blog object in db
     if not blog_in_db.first():
-        return {"error":f"Could not find blog with id {id}"}
+        return {"error": f"Could not find blog with id {id}"}
+    if not blog_in_db.first().author_id == author_id:
+        return {"error": "Only author can delete the blog"}
     blog_in_db.delete()
     db.commit()
     return {"msg":f"deleted blog with id {id}"}
